requirements.txt
Flask
python-dotenv
openai
werkzeug

